requirements.txt
django-ckeditor==6.4.2
django-crispy-forms
crispy-bootstrap5 